backend/pyproject.toml
[project]
name = "outreachai-backend"
version = "0.1.0"
description = "Python FastAPI backend for OutreachAI — LLM-powered patient outreach message generation"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "anthropic>=0.52",
    "google-genai>=1.14",
    "sse-starlette>=2.2",
]

[dependency-groups]
dev = [
    "ruff",
    "pytest",
    "pytest-asyncio",
    "httpx",
]
